+++ MPGES.LK/Core/Models/ContractStatusModel.swift
@@ -11,7 +11,7 @@ import UIKit
 public struct ContractStatusModel: Decodable {
     let id: Int
     let statusName: String
-    let value: Decimal
+    let value: Double
     
     enum CodingKeys: String, CodingKey {
         case id = "id"
